+++ src/components/Playground.js
@@ -79,7 +79,7 @@ const Playground = () => {
         options={[
           { label: "Option A", value: "a" },
           { label: "Option B", value: "b" },
-          { label: "Looooooooooooooooooooong option", value: "c" },
+          { label: "Loooooooooooooooooooooooooooooong option", value: "c" },
           { label: "Option d", value: "d" },
           { label: "Option e", value: "e" },
           { label: "Option f", value: "f" },